requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
pypdf>=4.0.0
pinecone-client>=5.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
